Implement winning, losing, and retrying
This commit should maybe have been broken into several smaller ones. To
add proper win/loss functionality, I made a bunch of significant
changes.

To have victory trigger properly, I had to modify the tile reveealing
functiton to return an integer, -1 in the case that the tile is a
mine, and otherwise the number of tiles cleared by the action. In the
main gameplay update loop, this value is used to update a struct field
representing the total number of unmined tiles that the player needs to
clear. When this number reaches 0, the round is considered to be won.
(For now this is hard coded in, but later this will all be
parameterized).

I broke the tea program update function into two smaller functions, one
for the main game loop, and one for the win/loss conditions.

I also used lipgloss to add a border around the field, and used figlet
to generate a litter header constant for the title.